Add support for detecting CRC instructions on Arm
