+++ source/arm/cpuid_arm.c
@@ -17,10 +17,47 @@
 #if (defined(_M_ARM) || defined(__arm__) || defined(__ARM_ARCH_ISA_A64))
 
 #    include <aws/checksums/private/cpuid.h>
+#    ifdef __linux__
+#        include <sys/auxv.h>
+#    endif
+
+static int32_t s_cpuid = 0;
 
 int aws_checksums_do_cpu_id(int32_t *cpuid) {
+#    ifdef __linux__
+    *cpuid = getauxval(AT_HWCAP);
+#    else
     (void)cpuid;
+#    endif
     return 0;
+}
+static void do_check(void) {
+    if (!s_cpuid) {
+        aws_checksums_do_cpu_id(&s_cpuid);
+    }
+}
+
+/** Returns non-zero if the CPU supports the PCLMULQDQ instruction. */
+int aws_checksums_is_clmul_present(void) {
+    return 0;
+}
+
+/** Returns non-zero if the CPU supports SSE4.1 instructions. */
+int aws_checksums_is_sse41_present(void) {
+    return 0;
+}
+
+/** Returns non-zero if the CPU supports SSE4.2 instructions (i.e. CRC32). */
+int aws_checksums_is_sse42_present(void) {
+    return 0;
+}
+
+/** Returns non-zero if the CPU support Arm CRC32/CRC32C instructions */
+int aws_checksums_is_arm_crc_present(void) {
+    const uint32_t hwcap_crc32 = (1 << 7);
+
+    do_check();
+    return s_cpuid & hwcap_crc32;
 }
 
 #endif
